Split server test cases into multiple files. #62
